backend/requirements.txt
fastapi
uvicorn[standard]
openai
langgraph
langchain-openai
langgraph-checkpoint-sqlite
aiosqlite
python-dotenv
pydantic
httpx
